requirements.txt
fastapi>=0.111.0
uvicorn>=0.29.0
databricks-sdk>=0.28.0
databricks-sql-connector>=3.0.0
openai>=1.30.0
pyarrow>=14.0.0
